<OC>llvm-gcc -O0/O1 -S -o 217_store_prod_O0/1.s 217_store_prod.c GCC实现128位整形的乘法细节:入参改为long类型

+++ LHiOSAccumulatesInObjectiveC/LHiOSAccumulatesInObjectiveC/LCTabs/LCDemo/LCDComputerSystem/LCDComputerSystemChapter3/cAndAssembleFile/217_store_prod.c
@@ -1,4 +1,4 @@
 typedef __int128 int128_t;
-void store_prod(int128_t *dest, int x, int y) {
+void store_prod(int128_t *dest, long x, long y) {
     *dest = x * (int128_t)y;
 }
